services/analytics-service/requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
SQLAlchemy==2.0.19
python-dotenv==1.0.0
requests==2.32.5



